KAT-10 refactor: 順位表画像 renderer を Core orchestrator + WMGP module layout に分離
