+++ apps/ledger/app/leagues/wmgp/league_module.rb
@@ -34,7 +34,7 @@ module Wmgp
       end
 
       def standings_table
-        ::StandingsExports::TableRenderer
+        ::Wmgp::Renderers::StandingsTable
       end
     end
